+++ packages/emoji/tests/useEmojiData.test.tsx
@@ -1,13 +1,18 @@
 import React from 'react';
+import { ShortcodesDataset } from 'emojibase';
 import { mockFetch, MockFetchResult, renderAndWait } from 'rut-dom';
 import { resetInstances } from '../src/EmojiDataManager';
 import { CanonicalEmoji, Source, UseEmojiDataOptions } from '../src/types';
 import useEmojiData, { resetLoaded } from '../src/useEmojiData';
 
-function cdn(locale: string, version: string = '1.0.0', compact: boolean = false) {
+function cdnEmojis(locale: string, version: string = '1.0.0', compact: boolean = false) {
   return `https://cdn.jsdelivr.net/npm/emojibase-data@${version}/${locale}/${
     compact ? 'compact' : 'data'
   }.json`;
+}
+
+function cdnShortcodes(preset: string, locale: string, version: string = '1.0.0') {
+  return `https://cdn.jsdelivr.net/npm/emojibase-data@${version}/${locale}/shortcodes/${preset}.json`;
 }
 
 describe('useEmojiData()', () => {
@@ -18,7 +23,11 @@ describe('useEmojiData()', () => {
   }
 
   function EmojiData(props: Props) {
-    const [emojis, source] = useEmojiData({ version: '1.0.0', ...props });
+    const [emojis, source] = useEmojiData({
+      shortcodes: ['emojibase'],
+      version: '1.0.0',
+      ...props,
+    });
 
     return <TestComp emojis={emojis} source={source} />;
   }
@@ -30,7 +39,6 @@ describe('useEmojiData()', () => {
       emoji: '😀',
       group: 0,
       hexcode: '1F600',
-      name: 'GRINNING FACE',
       order: 1,
       primary_shortcode: ':gleeful:',
       shortcodes: ['gleeful'],
@@ -44,19 +52,27 @@ describe('useEmojiData()', () => {
     },
   ];
 
+  const mockShortcodes: ShortcodesDataset = {
+    '1F600': ['gleeful'],
+  };
+
   let fetchSpy: MockFetchResult;
 
   beforeEach(() => {
     process.env.INTERWEAVE_ALLOW_FETCH_EMOJI = 'true';
 
     fetchSpy = mockFetch('/', 200)
-      .get(cdn('en', '1.0.0'), mockEmojis)
-      .get(cdn('en', '1.0.0', true), mockEmojis) // Compact
-      .get(cdn('it', '1.0.0'), mockEmojis)
-      .get(cdn('de', '2.0.0'), mockEmojis)
-      .get(cdn('ja', '1.2.3'), []) // No data
-      .get(cdn('fr', '1.0.0'), 404) // Not found
-      .get(cdn('fr', '2.0.0'), 404) // Not found
+      .get(cdnEmojis('en', '1.0.0'), mockEmojis)
+      .get(cdnEmojis('en', '1.0.0', true), mockEmojis) // Compact
+      .get(cdnEmojis('it', '1.0.0'), mockEmojis)
+      .get(cdnEmojis('de', '2.0.0'), mockEmojis)
+      .get(cdnEmojis('ja', '1.2.3'), []) // No data
+      .get(cdnEmojis('fr', '1.0.0'), 404) // Not found
+      .get(cdnEmojis('fr', '2.0.0'), 404) // Not found
+      .get(cdnShortcodes('emojibase', 'en', '1.0.0'), mockShortcodes)
+      .get(cdnShortcodes('emojibase', 'it', '1.0.0'), mockShortcodes)
+      .get(cdnShortcodes('emojibase', 'de', '2.0.0'), mockShortcodes)
+      .get(cdnShortcodes('emojibase', 'ja', '1.2.3'), mockShortcodes)
       .spy();
 
     // Emojibase caches requests
@@ -68,7 +84,8 @@ describe('useEmojiData()', () => {
   it('fetches data on mount', async () => {
     const result = await renderAndWait<Props>(<EmojiData />);
 
-    expect(fetchSpy.called(cdn('en'))).toBe(true);
+    expect(fetchSpy.called(cdnEmojis('en'))).toBe(true);
+    expect(fetchSpy.called(cdnShortcodes('emojibase', 'en'))).toBe(true);
 
     expect(result.root.findOne(TestComp)).toHaveProp('emojis', mockEmojis);
   });
@@ -80,7 +97,7 @@ describe('useEmojiData()', () => {
     await result.updateAndWait();
     await result.updateAndWait();
 
-    expect(fetchSpy.calls()).toHaveLength(1);
+    expect(fetchSpy.calls()).toHaveLength(2); // emojis + shortcodes
   });
 
   it('doesnt fetch if `avoidFetch` is true', async () => {
@@ -92,30 +109,34 @@ describe('useEmojiData()', () => {
   it('re-fetches data if a prop changes', async () => {
     const result = await renderAndWait<Props>(<EmojiData locale="ja" version="1.2.3" />);
 
-    expect(fetchSpy.called(cdn('ja', '1.2.3'))).toBe(true);
+    expect(fetchSpy.called(cdnEmojis('ja', '1.2.3'))).toBe(true);
+    expect(fetchSpy.called(cdnShortcodes('emojibase', 'ja', '1.2.3'))).toBe(true);
 
     await result.updateAndWait({
       locale: 'de',
       version: '2.0.0',
     });
 
-    expect(fetchSpy.called(cdn('de', '2.0.0'))).toBe(true);
+    expect(fetchSpy.called(cdnEmojis('de', '2.0.0'))).toBe(true);
+    expect(fetchSpy.called(cdnShortcodes('emojibase', 'de', '2.0.0'))).toBe(true);
   });
 
   it('supports compact datasets', async () => {
     await renderAndWait<Props>(<EmojiData compact />);
 
-    expect(fetchSpy.called(cdn('en', '1.0.0', true))).toBe(true);
+    expect(fetchSpy.called(cdnEmojis('en', '1.0.0', true))).toBe(true);
   });
 
   it('supports multiple locales', async () => {
     const result = await renderAndWait<Props>(<EmojiData locale="de" version="2.0.0" />);
 
-    expect(fetchSpy.called(cdn('de', '2.0.0'))).toBe(true);
+    expect(fetchSpy.called(cdnEmojis('de', '2.0.0'))).toBe(true);
+    expect(fetchSpy.called(cdnShortcodes('emojibase', 'de', '2.0.0'))).toBe(true);
 
     await result.rerenderAndWait(<EmojiData locale="it" />);
 
-    expect(fetchSpy.called(cdn('it'))).toBe(true);
+    expect(fetchSpy.called(cdnEmojis('it'))).toBe(true);
+    expect(fetchSpy.called(cdnShortcodes('emojibase', 'it', '1.0.0'))).toBe(true);
   });
 
   it('re-throws errors', async () => {